Add support for rotation
Add support for rotation of the csv node coordinates and corresponding motion directions. (Added in case we want to get motions for ESSI/OpenSees models oriented in different directions. This is more convenient to do in the converter code than creating separate ESSI/OpenSees models.) Will add rotation for the DRM nodes later (prefer to specify the rotation angle and reference point info also in the csv file as the DRM file and csv file are usually paired and thus use the same sets of settings)